+++ src/js/militarySymbolsObject.js
@@ -232,6 +232,48 @@ const militarySymbolsObject = {
       },
     },
   },
+  'Fly to Point (Normal)': {
+    type: 'Graphic Control Measure',
+    affiliation: {
+      friendly: {
+        path_1: {
+          d: 'm 60,45 80,0 m -40,55 -40,-55 0,-105 80,0 0,105 z',
+        },
+        text_1: {
+          symbolText: 'FTP',
+          x: '100',
+          y: '-20',
+          fontSize: '30',
+          textAnchor: 'middle',
+        },
+        text_2: {
+          symbolText: 'NRM',
+          x: '100',
+          y: '30',
+          fontSize: '30',
+          textAnchor: 'middle',
+        },
+      },
+      get friendlyTemplated() {
+        return this.friendly;
+      },
+      get hostile() {
+        return this.friendly;
+      },
+      get hostileTemplated() {
+        return this.friendly;
+      },
+      get unknown() {
+        return this.friendly;
+      },
+      get pending() {
+        return this.friendly;
+      },
+      get neutral() {
+        return this.friendly;
+      },
+    },
+  },
 };
 
 export default militarySymbolsObject;